Fixed debug build

+++ src/chips/ymfm_opn2.cpp
@@ -67,7 +67,7 @@ void YmFmOPN2::writeReg(uint32_t port, uint16_t addr, uint8_t data)
         m_headPos = 0;
 
     ++m_queueCount;
-    assert(m_queueCount < c_queueSize);
+    assert(m_queueCount < (long)c_queueSize);
 }
 
 void YmFmOPN2::writePan(uint16_t addr, uint8_t data)
